perf(osn-social): stop shipping passkey enrolment to every visitor
The security-events banner mounts on every settings visit and, until now,
pulled in `src/lib/webauthn.ts`, which held both the sign-in/step-up
ceremony and the passkey enrolment ceremony in one module. Because
`@simplewebauthn/browser` exports both through a single barrel with no
`sideEffects: false`, Rollup kept the barrel alive wherever either name was
imported and grouped `startAuthentication` and `startRegistration` into one
vendor chunk regardless of source layout, so the enrolment code rode along
into the entry bundle for visitors who never open the Security tab.

`webauthn.ts` is split into `webauthn-ceremony.ts` (assertion, used by the
banner and the Security tab alike) and `webauthn-registration.ts`
(attestation, used only by the Security tab). `vite.config.ts` marks the
`@simplewebauthn/browser` barrel side-effect-free for tree-shaking and
manually chunks its `startAuthentication` and `startRegistration` method
files apart, so the split holds through bundling instead of being undone by
the barrel. A production build now emits `webauthn-authentication` and
`webauthn-registration` as separate chunks, and neither appears in any
chunk the entry bundle loads eagerly — only the Security tab's existing
lazy import reaches the registration code.

Closes xchromo/osn-tracker#447
